# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Healthcare Platform API
  version: "1.0.0"
  description: >
    API for the Enterprise Healthcare Platform. This spec currently documents
    the implemented Auth and Patient modules in full; remaining modules are
    described at a summary level in API_DOCUMENTATION.md pending implementation.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
paths:
...
  /patients/{patientId}/lab-requests:
    get:
      summary: List a patient's lab requests, newest first
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Lab request list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabRequest'
  /patients/{patientId}/visits/{visitId}/lab-requests:
    post:
      summary: Request a lab test for a patient's visit (ADMIN, NURSE, DOCTOR)
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - in: path
          name: visitId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabRequestCreateRequest'
      responses:
        '201':
          description: Lab request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabRequest'
  /patients/{patientId}/lab-requests/{labRequestId}/status:
    put:
      summary: Update a lab request's workflow status (ADMIN, LAB_TECH)
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - in: path
          name: labRequestId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [REQUESTED, IN_PROGRESS, COMPLETED, CANCELLED]
      responses:
        '200':
          description: Updated lab request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabRequest'
        '409':
          description: Request already completed
  /patients/{patientId}/lab-requests/{labRequestId}/result:
    post:
      summary: Record a result against a lab request (ADMIN, LAB_TECH)
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - in: path
          name: labRequestId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resultData]
              properties:
                resultData:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Result recorded, request marked COMPLETED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabResult'
        '409':
          description: A result already exists for this request
    get:
      summary: Get the result recorded for a lab request
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - in: path
          name: labRequestId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lab result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabResult'
        '404':
          description: No result recorded yet
...
components:
  parameters:
    PatientId:
      in: path
      name: patientId
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LabRequestCreateRequest:
      type: object
      required: [testType]
      properties:
        testType:
          type: string
        notes:
          type: string
    LabRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        visitId:
          type: string
          format: uuid
        requestedById:
          type: string
          format: uuid
        testType:
          type: string
        status:
          type: string
          enum: [REQUESTED, IN_PROGRESS, COMPLETED, CANCELLED]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    LabResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        labRequestId:
          type: string
          format: uuid
        recordedById:
          type: string
          format: uuid
        resultData:
          type: object
          additionalProperties: true
        recordedAt:
          type: string
          format: date-time